ai_service/requirements.txt
agno>=2.5.3
fastapi[standard]>=0.115.8
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
httpx>=0.24.0
sqlite3-api>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sqlalchemy>=2.0.0
sqlalchemy-json>=0.7.0
duckduckgo-search>=6.0.0
